requirements.txt
streamlit
openai
scikit-learn
numpy
pandas
scipy
python-dotenv
